requirements.txt
numpy==1.16.4
sklearn-contrib-lightning==0.5.0
scikit-image==0.15.0
scikit-learn==0.21.2
torch==1.3.1
